feat: configuracion de perfiles 75%

+++ SGA_Api/SGA_Api/Models/Login/OperarioAlmacen.cs
@@ -2,7 +2,7 @@
 {
     public class OperarioAlmacen
     {
-        public int CodigoEmpresa { get; set; }
+        public short CodigoEmpresa { get; set; }
         public int Operario { get; set; }
         public string? CodigoAlmacen { get; set; }
     }
